backend/requirements.txt
asgiref==3.8.1
babel==2.17.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==44.0.2
defusedxml==0.7.1
Django==3.2
django-cleanup==9.0.0
django-filter==22.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djoser==2.2.0
flake8==7.1.2
gunicorn==23.0.0
idna==3.10
isort==6.0.1
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mccabe==0.7.0
oauthlib==3.2.2
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2025.1
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
